fix ci

+++ executor/hash_table_test.go
@@ -85,14 +85,14 @@ func (h hashCollision) BlockSize() int                    { panic("not implement
 func (s *pkgTestSerialSuite) TestHashRowContainer(c *C) {
 	hashFunc := fnv.New64
 	rowContainer, copiedRC := s.testHashRowContainer(c, hashFunc, false)
-	c.Assert(rowContainer.stat.probeCollision, Equals, 0)
+	c.Assert(rowContainer.stat.probeCollision, Equals, int64(0))
 	// On windows time.Now() is imprecise, the elapse time may equal 0
 	c.Assert(rowContainer.stat.buildTableElapse >= 0, IsTrue)
 	c.Assert(copiedRC.stat.probeCollision, Equals, rowContainer.stat.probeCollision)
 	c.Assert(copiedRC.stat.buildTableElapse, Equals, rowContainer.stat.buildTableElapse)
 
 	rowContainer, copiedRC = s.testHashRowContainer(c, hashFunc, true)
-	c.Assert(rowContainer.stat.probeCollision, Equals, 0)
+	c.Assert(rowContainer.stat.probeCollision, Equals, int64(0))
 	c.Assert(rowContainer.stat.buildTableElapse >= 0, IsTrue)
 	c.Assert(copiedRC.stat.probeCollision, Equals, rowContainer.stat.probeCollision)
 	c.Assert(copiedRC.stat.buildTableElapse, Equals, rowContainer.stat.buildTableElapse)
@@ -103,7 +103,7 @@ func (s *pkgTestSerialSuite) TestHashRowContainer(c *C) {
 	}
 	rowContainer, copiedRC = s.testHashRowContainer(c, hashFuncCollision, false)
 	c.Assert(h.count > 0, IsTrue)
-	c.Assert(rowContainer.stat.probeCollision > 0, IsTrue)
+	c.Assert(rowContainer.stat.probeCollision > int64(0), IsTrue)
 	c.Assert(rowContainer.stat.buildTableElapse >= 0, IsTrue)
 	c.Assert(copiedRC.stat.probeCollision, Equals, rowContainer.stat.probeCollision)
 	c.Assert(copiedRC.stat.buildTableElapse, Equals, rowContainer.stat.buildTableElapse)
